@@ -1,4 +1,4 @@
-﻿using System;
+﻿using System.Linq;
 using Task3.DoNotChange;
 
 namespace Task3
@@ -7,29 +7,16 @@ namespace Task3
     {
         private readonly IUserDao _userDao;
 
-        public UserTaskService(IUserDao userDao)
-        {
-            _userDao = userDao;
-        }
+        public UserTaskService(IUserDao userDao) => _userDao = userDao;
 
         public int AddTaskForUser(int userId, UserTask task)
         {
-            if (userId < 0)
-                return -1;
-
+            UserErrors.CheckUser(userId);
             var user = _userDao.GetUser(userId);
-            if (user == null)
-                return -2;
-
+            UserErrors.CheckUser(user);
             var tasks = user.Tasks;
-            foreach (var t in tasks)
-            {
-                if (string.Equals(task.Description, t.Description, StringComparison.OrdinalIgnoreCase))
-                    return -3;
-            }
-
+            tasks.ToList().ForEach(t => UserErrors.CheckTasks(t.Description, task.Description));
             tasks.Add(task);
-
             return 0;
         }
     }
